Refactor app architecture, implement DI using Hilt

+++ app/src/main/java/com/spksh/todoline/data/Converters.kt
@@ -5,12 +5,12 @@ import java.time.Instant
 
 class Converters {
     @TypeConverter
-    fun fromList(list: List<Int>): String {
+    fun fromList(list: List<Long>): String {
         return list.joinToString(",")
     }
 
     @TypeConverter
-    fun toList(data: String): List<Int> {
-        return if (data.isEmpty()) emptyList() else data.split(",").map { it.toInt() }
+    fun toList(data: String): List<Long> {
+        return if (data.isEmpty()) emptyList() else data.split(",").map { it.toLong() }
     }
 }
