+++ app/src/main/java/com/spksh/todoline/data/Task.kt
@@ -5,15 +5,15 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "task_table")
 data class Task(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String = "",
     val description: String = "",
     val importance: Int = 10,
     val urgency: Int = 10,
     val deadline: Long? = null,
     val requiredTime: Int = 0,
-    val tagsIds: List<Int> = emptyList(),
-    val parentTaskId: Int? = null,
-    val childTasksIds: List<Int> = emptyList(),
+    val tagsIds: List<Long> = emptyList(),
+    val parentTaskId: Long? = null,
+    val childTasksIds: List<Long> = emptyList(),
     val progress: Float = 0f,
 )
